Разделил  main.js на отдельные модули
