+++ js/data.js
@@ -1,0 +1,36 @@
+const NAMES = [
+  'Виктор', 'София', 'Роман', 'Алиса', 'Константин',
+  'Дарья', 'Никита', 'Полина', 'Максим', 'Вероника',
+  'Георгий', 'Ксения', 'Арсений', 'Анастасия', 'Тимофей'
+];
+
+// Список комментариев
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+// Список описаний для фотографий
+const DESCRIPTIONS = [
+  'Рассвет в горах',
+  'Вечер у камина',
+  'Прогулка по парку',
+  'Семейное застолье',
+  'Тихое утро на берегу озера',
+  'Яркие краски летнего фестиваля',
+  'Приключения в далеких странах',
+  'Щенок в саду',
+  'Вечерний город с высоты',
+  'Зима в лесу',
+  'В ресторане с друзьями',
+  'Спортивные достижения и победы',
+  'Романтический ужин при свечах',
+  'Творческий процесс в мастерской',
+  'Пикник на природе'
+];
+
+export { NAMES, MESSAGES, DESCRIPTIONS };
